Edited layout component and layout of pages
Created a better `Layout` component
  * Allows for props to be passed, such as `home` (bool) to either put a hero img or not
Fixed a bug with the `Hero` component
Used the new layout method for all of the other pages
Created `Hero.module.css` and added a style to fix the size of the header (though still needs work)
  -> Suggest a new convention: use uppercase names for css module files for components, and lowercase for non-components
Create `utils.module.css` and added a style to put an element behind other elements (used for the image in the `Hero` component)
  * This file will be used for utility classes, like `roundedBox`, or `boxShadow`, etc.

No errors

Still need to work on the CSS and styles for the `Hero` component to stop the overlap.
Also, a general style-set should be developed (ie. accent colors, bg-colors)
  * These colors would be placeholders until the commission gives us the theme they developed