Events, Fonts, and Compiler Fix Attempts
When trying to create boxes for events, compiler did not  allow @next/font. So, I  ran  "npm install @next/font" and that still didnt work. So, I added something in next config to force compile with SWC instead  of babel but that   could cause app crashing so I commented  it.